+++ projects/gloo/pkg/syncer/setup_syncer_test.go
@@ -136,31 +136,6 @@ var _ = Describe("SetupSyncer", func() {
 
 		})
 	})
-
-	Context("WaitForReady", func() {
-		It("should timeout when channels don't warm up", func() {
-			timeout := time.Millisecond
-			err := WaitForReady(ctx, timeout, make(chan struct{}))
-			Expect(err).To(MatchError(context.DeadlineExceeded))
-		})
-		It("should succeed when channels are ready", func() {
-			timeout := time.Millisecond
-			closedChannels0, closedChannels1 := make(chan struct{}), make(chan struct{})
-			close(closedChannels0)
-			close(closedChannels1)
-			err := WaitForReady(ctx, timeout, closedChannels0, closedChannels1)
-			Expect(err).NotTo(HaveOccurred())
-		})
-		It("should exit when context is canceled", func() {
-			timeout := time.Second
-			go func() {
-				time.Sleep(time.Second / 3)
-				cancel()
-			}()
-			err := WaitForReady(ctx, timeout, make(chan struct{}))
-			Expect(err).To(MatchError(context.Canceled))
-		})
-	})
 })
 
 func getRandomAddr() string {
